Store long value from parse
git-svn-id: https://threeten.svn.sourceforge.net/svnroot/threeten/trunk/threeten@1383 291d795c-afe8-5c46-8ee5-bf9dd72e1864

+++ src/main/java/javax/time/calendar/format/ReducedPrinterParser.java
@@ -96,7 +96,7 @@ final class ReducedPrinterParser extends NumberPrinterParser {
         if (value < baseValue) {
             value += range;
         }
-        context.setParsedField(rule, (int) value);
+        context.setParsedField(rule, value);
     }
 
     //-----------------------------------------------------------------------
